pyproject.toml
[project]
name = "dcase-2026"
version = "0.1.0"
description = "DCASE 2026"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.4",
    "easydict>=1.13",
    "einops>=0.8.2",
    "hear21passt>=0.0.26",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "isort>=8.0.1",
    "librosa>=0.11.0",
    "mamba-ssm>=2.3.1",
    "matplotlib>=3.10.8",
    "nbstripout>=0.9.1",
    "nltk>=3.9.4",
    "noisereduce>=3.0.3",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pre-commit>=4.5.1",
    "py7zr>=1.1.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.3.0",
    "soundfile>=0.13.1",
    "timm>=1.0.26",
    "torchaudio>=2.11.0",
    "torchcodec>=0.11.0",
    "torchlibrosa>=0.1.0",
    "torchmetrics>=1.9.0",
    "transformers>=5.5.0",
    "webdataset>=1.0.2",
]
